+++ utils/tests/conttest.cc
@@ -7,7 +7,7 @@
 
 static void *nonfunc(void *_x)
 {
-    int x = (int)_x;
+    long x = (long)_x;
     
     return (void *)(1234560000 + x);
 }
@@ -15,7 +15,7 @@ static void *nonfunc(void *_x)
 
 static void *func(void *_x)
 {
-    int x = (int)_x;
+    long x = (long)_x;
     
     for (int count = 0; count < 4 && WvCont::isok(); count++)
 	WvCont::yield((void *)++x);
@@ -41,7 +41,7 @@ public:
 private:
     void *honker(Honk &h, void *_x)
     {
-	int x = (int)_x;
+	long x = (long)_x;
 	printf("%s: STARTING (%d)\n", id, x);
 	
 	for (x--; WvCont::isok() && x > 0; x--)
@@ -71,21 +71,21 @@ int main()
 	// 'func' actually gets called; there are no parallel running 'func's.
 	// cb2's task calls into cb1's task, however.
     
-	printf("zot1: %d\n", (int)cb1((void *)100));
-	printf("zot1: %d\n", (int)cb2((void *)200));
-	printf("zot1: %d\n", (int)cb3((void *)300));
+	printf("zot1: %d\n", (long)cb1((void *)100));
+	printf("zot1: %d\n", (long)cb2((void *)200));
+	printf("zot1: %d\n", (long)cb3((void *)300));
 	cb1 = WvCont(nonfunc);
-	printf("zot2: %d\n", (int)cb1((void *)400));
-	printf("zot2: %d\n", (int)cb2((void *)500));
-	printf("zot2: %d\n", (int)cb3((void *)600));
+	printf("zot2: %d\n", (long)cb1((void *)400));
+	printf("zot2: %d\n", (long)cb2((void *)500));
+	printf("zot2: %d\n", (long)cb3((void *)600));
 	cb2 = nonfunc;
-	printf("zot3: %d\n", (int)cb1((void *)700));
-	printf("zot3: %d\n", (int)cb2((void *)800));
-	printf("zot3: %d\n", (int)cb3((void *)900));
+	printf("zot3: %d\n", (long)cb1((void *)700));
+	printf("zot3: %d\n", (long)cb2((void *)800));
+	printf("zot3: %d\n", (long)cb3((void *)900));
 	cb3 = nonfunc;
-	printf("zot4: %d\n", (int)cb1((void *)1000));
-	printf("zot4: %d\n", (int)cb2((void *)1100));
-	printf("zot4: %d\n", (int)cb3((void *)1200));
+	printf("zot4: %d\n", (long)cb1((void *)1000));
+	printf("zot4: %d\n", (long)cb2((void *)1100));
+	printf("zot4: %d\n", (long)cb3((void *)1200));
     }
     
     // fun with recursive continuations.  If this doesn't do something
